requirements.txt
Django>=4.2
mysqlclient
python-dotenv
django-widget-tweaks
